test: keep marker guard deterministic and pin test file locations
Review on #210: the deployment-marker required-field check is fully
deterministic and was sitting in test/live/, so npm test could no longer
fail it. Move it back to test/schema.test.ts and share the endpoint
triples via test/helpers/schema-endpoints.ts.

Also assert that every .test.ts is directly under test/ or test/live/,
because the narrowed globs would otherwise green-pass a misfiled file.
Fixes the live-probe-gate comment typo.
