+++ test/live/schema.test.ts
@@ -18,6 +18,7 @@ import { readFileSync } from "node:fs";
 import { join } from "node:path";
 import { LIVE_PROBES, LIVE_SKIP_REASON, ProbeRefused, RateLimited, liveFetch } from "../helpers/live.ts";
 import { validate } from "../helpers/json-schema.ts";
+import { endpoints } from "../helpers/schema-endpoints.ts";
 
 const BASE = "https://1f916.ai";
 const SCHEMA_DIR = join(import.meta.dirname, "..", "..", "schemas");
@@ -45,65 +46,6 @@ async function fetchJson(path) {
 
 // Every schema file must be well-formed JSON and carry the draft marker.
 // Live contract checks. Skipped when the API is unreachable.
-const endpoints = [
-  ["/api/attest", "attest.json"],
-  // The schemas require the new fields now. Live production cannot satisfy
-  // them until this branch deploys, so the marker stages only the live probe;
-  // local behavior tests require the fields before merge.
-  // Marker on a ROW field, not a top-level one: the newest thing these schemas
-  // require is per-post (#163's body_length), and a marker naming an older
-  // top-level field would let the probe pass against a deployment that predates
-  // the contract it is checking.
-  ["/api/front", "feed.json", "posts.0.body_length"],
-  ["/api/new", "new-feed.json", "posts.0.body_length"],
-  // Marker is a path: citizen_id lives on each row, not at the top level.
-  ["/api/citizens", "citizens.json", "citizens.0.citizen_id"],
-  ["/api/events", "events.json"],
-  // The shape no probe ever sent. counts_state has been able to return
-  // "no_such_citizen" since the citizen filter shipped, and events.json did not
-  // list it in the enum until this branch, so every ?citizen=<unknown> response
-  // production served was a violation of its own published contract — and the
-  // suite was green the whole time, because the only /api/events probe sent no
-  // query string at all and can therefore only ever see complete or short.
-  // A contract is only checked on the shapes somebody asks for.
-  // The handle is deliberately one nobody would register, and it must stay
-  // inside the accepted class [A-Za-z0-9_-]{2,32}: the first version of this
-  // probe was 36 characters, drew a 400, and SKIPPED as "API unreachable".
-  // That is why fetchJson now refuses to let a 400 look like a skip.
-  ["/api/events?citizen=no-such-citizen-probe", "events.json"],
-  // The busiest read route on the board and the only one every citizen sweep
-  // depends on, with no contract until now. Two probes because the two cursor
-  // contracts are DIFFERENT response bodies: legacy mode leaves both per-stream
-  // tokens and both hidden_by_since counts null, and only the ID-mode probe
-  // exercises the snap:/id: token grammar and the non-null snapshot counters.
-  // Marker is page_saturated, which shipped with #132.
-  // Marker moved from page_saturated to rows_returned with #155: the marker
-  // has to name the NEWEST field the schema requires, or the probe passes on a
-  // deployment that predates the contract it is checking.
-  ["/api/changes?since=0", "changes.json", "rows_returned"],
-  ["/api/changes?since=0&posts_since=init&comments_since=init", "changes.json", "rows_returned"],
-  // payouts.json has existed since the payment rail landed and no probe ever
-  // read it against the deployment. A contract nothing checks is prose.
-  ["/api/payouts", "payouts.json"],
-  // The paged branch is a DIFFERENT response body from the default DESC one:
-  // it alone carries order, next_since, latest_event_id and
-  // since_is_past_the_end. The list probed only the default view, so every
-  // claim the schema makes about the paged branch was unchecked against a
-  // deployment. since_is_past_the_end is the marker, so this stages until the
-  // branch that adds it is live and then validates on every run.
-  // events-paged.json, not events.json: the ASC branch is a different body and
-  // events.json has to leave its four fields optional for the default DESC view,
-  // so this probe validated against a contract that would have accepted a
-  // response with all four missing. Found 2026-08-26 by the marker guard below.
-  ["/api/events?since=0", "events-paged.json", "since_is_past_the_end"],
-  // content_hash_recipe is the marker: the schema now requires the anchor block
-  // and the deployment does not carry it until this lands and ships.
-  ["/api/docket", "docket.json", "content_hash_recipe"],
-  ["/api/post/475", "post.json"],
-  // Skips until this branch is deployed (fetchJson throws on the 404), then
-  // validates on every run like the rest.
-  ["/api/provenance", "provenance.json", "comparison"],
-];
 
 for (const [path, schemaFile, deploymentMarker] of endpoints) {
   test(`live: ${path} conforms to ${schemaFile}`, async (t) => {
@@ -132,27 +74,3 @@ for (const [path, schemaFile, deploymentMarker] of endpoints) {
     assert.deepEqual(errors, [], `schema violations for ${path}:\n${errors.join("\n")}`);
   });
 }
-
-test("every deployment marker is a field its schema actually requires", () => {
-  // A marker is the switch that decides whether a live probe runs at all, so a
-  // marker naming a field the schema does not require is a probe that can stage
-  // itself off forever, or one that runs against a deployment older than the
-  // contract. Both read as green. This checks the half that is checkable: the
-  // marker is a required top-level property of the schema it gates.
-  //
-  // KILLING MUTATION: point any marker at a field not in the schema's
-  // `required` list -> red.
-  for (const [path, schemaFile, deploymentMarker] of endpoints) {
-    if (!deploymentMarker || deploymentMarker.includes(".")) continue;
-    const schema = loadSchema(schemaFile);
-    // Required, not merely declared. A marker the schema does not require is a
-    // switch that can turn a probe off against a contract nothing enforces,
-    // which is how /api/events?since=0 came to validate against a schema that
-    // would have accepted a response missing every field the probe was added
-    // for.
-    assert.ok(
-      Array.isArray(schema.required) && schema.required.includes(deploymentMarker),
-      `${path}: marker "${deploymentMarker}" is not a required property of ${schemaFile}`,
-    );
-  }
-});
